Timesheet Charge Role functionallity completed.
Former-commit-id: a51f6066a028fc8abd0e226bef3999469c1deb17

+++ acm-forms/acm-form-time/src/main/java/com/armedia/acm/form/time/model/TimeItem.java
@@ -23,7 +23,7 @@ public class TimeItem
 
     private List<String> chargeRoles;
 
-    private Long totalCost;
+    private Double totalCost;
 
     private String title;
 
@@ -111,11 +111,11 @@ public class TimeItem
     }
 
     @XmlElement(name = "totalCost")
-    public Long getTotalCost() {
+    public Double getTotalCost() {
         return totalCost;
     }
 
-    public void setTotalCost(Long totalCost) {
+    public void setTotalCost(Double totalCost) {
         this.totalCost = totalCost;
     }
